Move appendToDom, removeFromDom functions to module handleDom.js
