+++ assets/javascript/index.js
@@ -3,10 +3,9 @@ import {
   listHandler, addBookHandler, contactHandler,
 } from './modules/nav.js';
 import setDateTime from './modules/setDateTime.js';
-import { retrieveData } from './modules/handleData.js';
 import Book from './modules/book.js';
+import { loadBooks, appendBookToDOM } from './modules/handleDOM.js';
 
-const booksList = document.getElementById('books_list');
 const bookForm = document.getElementById('booksForm');
 
 export const contactSection = document.getElementById('contact-section');
@@ -14,34 +13,6 @@ export const addBookSection = document.getElementById('form_section');
 export const booksSection = document.getElementById('books_section');
 
 export const dateElement = document.getElementById('date');
-
-let booksArray = [];
-
-const removeBookFromDOM = (book) => {
-  const removeBtn = document.getElementById(book.id);
-  const bookItem = removeBtn.parentElement;
-  bookItem.parentElement.removeChild(bookItem);
-};
-
-const appendBookToDOM = (book) => {
-  const bookItem = document.createElement('li');
-  const removeButton = document.createElement('button');
-  bookItem.classList.add('book_item');
-  removeButton.innerText = 'Remove';
-  removeButton.setAttribute('type', 'button');
-  removeButton.setAttribute('id', book.id);
-  removeButton.classList.add('remove');
-  bookItem.innerHTML = `
-    <span>"${book.title}"   by   ${book.author}</span> 
-   
-    `;
-  bookItem.append(removeButton);
-  booksList.append(bookItem);
-  removeButton.addEventListener('click', () => {
-    Book.removeBook(book);
-    removeBookFromDOM(book);
-  });
-};
 
 bookForm.addEventListener('submit', (event) => {
   event.preventDefault();
@@ -54,15 +25,8 @@ bookForm.addEventListener('submit', (event) => {
   appendBookToDOM(newBook);
 });
 
-function loadBooks() {
-  booksArray.forEach((book) => {
-    appendBookToDOM(book);
-  });
-}
-
 window.onload = () => {
   booksSection.classList.add('visible');
-  booksArray = retrieveData();
   loadBooks();
   setDateTime();
 };
